pyproject.toml
[project]
name = "debate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "langchain>=1.2.9",
    "ollama>=0.6.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.21",
]
